refactor: 테스트의 FakeNumberGenerator를 재사용성을 위해 util로 분리
